Modify header layout
Changed header layout to have four separate sections for the content to
go into in order to control each part individually. Added padding and
a shadow in order to style it like the prompt. Adjusted the first
section to display as a flex container in order to have the search icon
right next to the search bar