Modify header style
Removed line break from html in order to style that section better in
css. Added more selectors in css in order to style each section of the
header more specifically. Added styling to the various sections in the
header in order to make the header look more like the prompt. Used
flexbox properties on sub-containers in the header for easier styling
of items closely related to each other