Modify announcements and trending sections
Added elements to announcements and trending sections in order to setup
the layout for them. Added id's, classes, and selectors to the
announcements and trending sections in order to better target them for
styling. Styled announcements and trending sections with both grid and
flexbox properties in order to make them look more like the prompt