Add icons and pfp's
Added icons folder to hold all of the icons and pfp's for the page.
Changed the html so that the icon images and pfp images have their
respective classes for styling. Adjusted classes of icons and pfp's
in css so that they would look more like the prompt. Adjusted sidebar
styling so that everything lines up